requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-login==0.6.3
docker==6.1.3
python-dotenv==1.0.0
paramiko==3.3.1
psutil==5.9.6
gunicorn==21.2.0
eventlet==0.33.3
gevent==22.10.2
gevent-websocket==0.10.1
Werkzeug==3.0.1
